homework/hw01/requirements.txt
matplotlib==3.7.0
numpy==1.23.5
Pillow==9.5.0
python-dotenv==1.0.0
